+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 
+- Fix payment method funding and deletion failures due to undetermined payment method type
 - Adds `refund` function in Insurance service for requesting a refund for a standalone insurance
 
 ## v7.1.0 (2024-01-08)